backend/requirements.txt
aiosmtplib==2.0.1
alembic==1.9.1
amqp==5.1.1
asgiref==3.6.0
asttokens==2.4.0
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.4
backcall==0.2.0
base36==0.1.1
bcrypt==4.0.1
billiard==3.6.4.0
cachetools==5.3.1
celery==5.2.7
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
cffi==1.15.1
charset-normalizer==2.1.1
click @ file:///C:/ci/click_1646056762388/work
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama @ file:///C:/Windows/TEMP/abs_9439aeb1-0254-449a-96f7-33ab5eb17fc8apleb4yn/croots/recipe/colorama_1657009099097/work
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.2.35
cryptography==41.0.1
Cython==3.0.11
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
dj-database-url==1.3.0
Django==4.2.6
django-allauth==0.54.0
django-celery-beat==2.5.0
django-celery-results==2.5.0
django-cors-headers==3.14.0
django-debug-toolbar==4.2.0
django-email-verification==0.3.1
django-environ==0.11.2
django-filter==23.2
django-heroku==0.3.1
django-oauth-toolkit==3.0.1
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
django-timezone-field==5.0
Django-Verify-Email==2.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
dnspython==2.2.1
drf-spectacular==0.26.4
drf-yasg==1.21.7
email-validator==1.3.0
exceptiongroup==1.1.3
executing==2.0.0
Flask @ file:///C:/b/abs_ef16l83sif/croot/flask_1671217367534/work
flask-abort==0.0.1
Flask-Bcrypt==1.0.1
Flask-Dance==6.2.0
Flask-Login==0.6.2
Flask-Migrate==4.0.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
glcontext==3.0.0
google-api-python-client==1.6.4
google-auth==2.22.0
google-auth-oauthlib==1.0.0
greenlet==2.0.1
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
inflection==0.5.1
ipdb==0.13.13
ipython==8.16.0
itsdangerous @ file:///tmp/build/80754af9/itsdangerous_1621432558163/work
itypes==1.2.0
jedi==0.19.0
Jinja2 @ file:///C:/b/abs_7cdis66kl9/croot/jinja2_1666908141852/work
jsonpickle==0.9.6
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jwcrypto==1.5.6
kombu==5.2.4
Mako==1.2.4
Markdown==3.4.3
MarkupSafe @ file:///C:/ci/markupsafe_1654508036328/work
matplotlib-inline==0.1.6
moderngl==5.12.0
nbtlib==2.0.4
node==1.2
nodejs==0.1.1
numpy==1.24.1
oauthlib==3.2.2
odict==1.9.0
openapi-codec==1.3.2
optional-django==0.1.0
packaging==23.1
pandas==1.5.3
parso==0.8.3
passlib==1.7.4
pickleshare==0.7.5
Pillow==9.3.0
plumber==1.7
prompt-toolkit==3.0.38
psycopg2==2.9.5
psycopg2-binary==2.9.10
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
pygame==2.6.1
pyglet==2.1.0
PyGLM==2.7.3
Pygments==2.16.1
PyJWT==2.7.0
pyparsing==3.1.1
python-crontab==2.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7.1
PyWavefront==1.3.3
PyYAML==6.0.1
redis==4.5.4
referencing==0.30.2
requests==2.28.1
requests-oauthlib==1.3.1
rpds-py==0.10.3
rsa==4.9
simplejson==3.19.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
SQLAlchemy==1.4.45
sqlparse==0.4.3
stack-data==0.6.3
tomli==2.0.1
traitlets==5.10.1
typing_extensions==4.5.0
tzdata==2022.7
uritemplate==3.0.1
urllib3==1.26.13
URLObject==2.4.3
vine==5.0.0
wcwidth==0.2.6
Werkzeug @ file:///C:/b/abs_17q5kgb8bo/croot/werkzeug_1671216014857/work
whitenoise==6.4.0
wincertstore==0.2
WTForms==3.0.1
zope.component==5.1.0
zope.deferredimport==4.4
zope.deprecation==4.4.0
zope.event==4.6
zope.hookable==5.4
zope.interface==5.5.2
zope.lifecycleevent==4.4
zope.proxy==5.0.0
